pyproject.toml
[tool.poetry]
name = "ci-cd-example"
version = "0.1.0"
description = ""
authors = ["paulcjh <ph@mystic.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^24.3.0"
pipeline-ai = "^2.1.3"
fastapi = "^0.110.1"
pre-commit = "^3.7.0"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
